backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
mysql-connector-python==8.0.26
PyJWT==2.1.0
requests==2.26.0
werkzeug==2.0.1
boto3==1.18.65
alibabacloud-core==1.0.1
alibabacloud-ecs==1.0.0
tencentcloud-sdk-python==3.0.578
huaweicloudsdkcore==3.1.15
huaweicloudsdkecs==3.1.15
azure-identity==1.7.1
azure-mgmt-compute==23.1.0
azure-mgmt-resource==20.0.0
azure-mgmt-network==19.1.0
google-cloud-compute==0.5.0
google-cloud-storage==1.42.3
google-auth==2.3.0
volcenginesdk==1.0.23
python-dotenv==0.19.1
jsonschema==4.2.1
cachetools==4.2.4
openai>=1.6.0
fsspec>=2023.1.0
Pillow>=10.0.0
tiktoken>=0.5.0
